requirements.txt
PyQt5==5.15.9
opencv-python==4.8.1.78
face-recognition==1.3.5
dlib==19.24.4
pyttsx3==2.90
sounddevice==0.4.6
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
Pillow==10.0.0
requests==2.31.0
